use std::string as buffer
