+++ 10_bulk_server/async.hpp
@@ -1,13 +1,14 @@
 #pragma once
 
 #include <cstddef>
+#include <string>
 
 namespace async {
 
     using handle_t = void*;
 
     handle_t connect(std::size_t bulk);
-    void receive(handle_t handle, const char* data, std::size_t size);
+    void receive(handle_t handle, std::string&& input);
     void disconnect(handle_t handle);
 
 } //namespace async
